requirements.txt
python-dotenv
langchain
openai
slack-sdk
slack-bolt
flask
azure-core
datetime